Version 2.12i
Version 2.12i

+++ src/main/java/nc/util/EnergyHelper.java
@@ -5,7 +5,7 @@ import nc.config.NCConfig;
 public class EnergyHelper {
 	
 	public static int getEUTier(double powerRF) {
-		double euPerTick = (double) powerRF / (double) NCConfig.rf_per_eu;
+		double euPerTick = powerRF / (double) NCConfig.rf_per_eu;
 		for (int i = 1; i < 10; i++) if (euPerTick <= Math.pow(2, 2*i + 3)) return i;
 		return 10;
 	}
